requirements.txt
dj-database-url==0.5.0
Django==2.1.7
django-heroku==0.3.1
djangorestframework==3.9.2
gunicorn==19.9.0
mysqlclient==1.4.2.post1
psycopg2==2.7.7
PyMySQL==0.9.3
pytz==2018.9
virtualenv==16.4.3
whitenoise==4.1.2
